+++ .eslintrc.json
@@ -1,1 +1,1 @@
-{  "extends": [    "airbnb",    "prettier",    "plugin:import/errors",    "plugin:import/warnings",    "plugin:import/typescript",    "plugin:@typescript-eslint/recommended"  ],  "plugins": ["prettier", "@typescript-eslint"],  "parserOptions": {    "sourceType": "module"  },  "rules": {    "prettier/prettier": [      "error",      {        "endOfLine": "auto"      }    ],    "react/jsx-filename-extension": [1, { "extensions": [".tsx", ".ts"] }],    "react/react-in-jsx-scope": "off",    "import/extensions": [      "error",      "ignorePackages",      {        "js": "never",        "jsx": "never",        "ts": "never",        "tsx": "never"      }    ],    "no-use-before-define": "off",    "no-unused-vars": "off",    "no-console": "off",    "func-names": "off",    "no-process-exit": "off",    "object-shorthand": "off",    "class-methods-use-this": "off"  }}
+{  "extends": [    "airbnb",    "prettier",    "plugin:import/errors",    "plugin:import/warnings",    "plugin:import/typescript",    "plugin:@typescript-eslint/recommended"  ],  "plugins": ["prettier", "@typescript-eslint"],  "parserOptions": {    "sourceType": "module"  },  "rules": {    "prettier/prettier": [      "error",      {        "endOfLine": "auto"      }    ],    "no-shadow": "off",    "@typescript-eslint/no-shadow": ["error"],    "react/jsx-filename-extension": [1, { "extensions": [".tsx", ".ts"] }],    "react/react-in-jsx-scope": "off",    "import/extensions": [      "error",      "ignorePackages",      {        "js": "never",        "jsx": "never",        "ts": "never",        "tsx": "never"      }    ],    "no-use-before-define": "off",    "no-unused-vars": "off",    "no-console": "off",    "func-names": "off",    "no-process-exit": "off",    "object-shorthand": "off",    "class-methods-use-this": "off"  }}
