+++ package.json
@@ -24,7 +24,6 @@
     "react-firebase-hooks": "^3.0.4",
     "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
-    "react-router-redux": "^4.0.8",
     "react-scripts": "4.0.3",
     "redux": "^4.1.0",
     "redux-actions": "^2.6.5",
