+++ src/pages/Board/consts/consts.ts
@@ -1,0 +1,34 @@
+export const addColumnContent = {
+  content: "Enter a name for the column and click the Add button",
+  title: "Add column",
+  label: "Column name",
+  button: "Add",
+};
+
+export const addItemContent = {
+  content: "Enter a text for the item and click the Add button",
+  title: "Add Item",
+  label: "Item text",
+  button: "Add",
+};
+
+export const editBoardNameContent = {
+  content: "Enter a name for the board and click the Edit button",
+  title: "Edit board name",
+  label: "Board name",
+  button: "Edit",
+};
+
+export const editColumnNameContent = {
+  content: "Enter a name for the column and click the Edit button",
+  title: "Edit column name",
+  label: "Column name",
+  button: "Edit",
+};
+
+export const editItemContent = {
+  content: "Enter a text for the item and click the Edit button",
+  title: "Edit item text",
+  label: "Item text",
+  button: "Edit",
+};
